enhanced_requirements.txt
# Enhanced RAG System Requirements
# Core dependencies from original system
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv
typesense
langchain_openai
langgraph

# Web interface dependencies
streamlit>=1.28.0
streamlit-chat>=0.1.1
flask>=2.3.0

# Additional dependencies for enhanced features
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
tqdm>=4.62.0

# Optional: For advanced text processing
nltk>=3.7
spacy>=3.4.0

# Optional: For better document parsing
pdfplumber>=0.7.0
python-docx>=0.8.11
openpyxl>=3.0.9

# Optional: For visualization and analysis
plotly>=5.0.0
wordcloud>=1.8.0